Extract startup logic to separate class
